fixed flujo de registro con la base de datos
