+++ database/index.ts
@@ -1,2 +1,3 @@
 export * as db from './db';
 export * as dbUsers from './dbUsers';
+export * as dbUsersWebAuthn from './dbUsersWebAuthn';
